# portal/src/main/java/com/apteka/portal/controllers/GroupMainPageLinkController.java
package com.apteka.portal.controllers;

import java.util.List;

import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.security.core.annotation.AuthenticationPrincipal;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PatchMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

import com.apteka.portal.dtos.request.mainpagelinks.GroupMainPageLinkUpdateRequestDTO;
import com.apteka.portal.dtos.request.mainpagelinks.GroupMainPageLinksRequestDTO;
import com.apteka.portal.dtos.response.mainpagelink.GroupMainPageLinksResponseDTO;
import com.apteka.portal.models.AppUserDetails;
import com.apteka.portal.services.GroupMainPageLinksService;

import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;

@RestController
@RequestMapping("/api/v1/groups-main-page-links")
@RequiredArgsConstructor
public class GroupMainPageLinkController {
    private final GroupMainPageLinksService groupMainPageLinksService;

    @PreAuthorize("""
        @security.hasAction('CAN_SELECT_NON_ACTIVE_MAIN_PAGE_LINK')
        or @security.hasAction('CREATE_MAIN_PAGE_LINK')
        or @security.hasAction('UPDATE_MAIN_PAGE_LINK')
        or @security.hasAction('SAFE_DELETE_MAIN_PAGE_LINK')
        or @security.hasAction('CAN_ACTIVATE_MAIN_PAGE_LINK_AFTER_SAFE_DELETE')
        or @security.hasAction('PERMANENT_DELETE_MAIN_PAGE_LINK')
        or @security.hasRole('ADMIN')
        """)
    @GetMapping("/get-all")
    public ResponseEntity<List<GroupMainPageLinksResponseDTO>> getAll(
            @RequestParam(defaultValue = "true") Boolean isActive,
            @AuthenticationPrincipal AppUserDetails currentUser) {
        return ResponseEntity.ok(groupMainPageLinksService.getAll(isActive, currentUser));
    }

    @PreAuthorize("@security.hasAction('CREATE_MAIN_PAGE_LINK') or @security.hasRole('ADMIN')")
    @PostMapping
    public ResponseEntity<GroupMainPageLinksResponseDTO> create(@Valid @RequestBody GroupMainPageLinksRequestDTO dto) {
        return ResponseEntity.status(HttpStatus.CREATED).body(groupMainPageLinksService.create(dto));
    }

    @PreAuthorize("@security.hasAction('UPDATE_MAIN_PAGE_LINK') or @security.hasRole('ADMIN')")
    @PutMapping("/{id}")
    public ResponseEntity<GroupMainPageLinksResponseDTO> update(@PathVariable Integer id,
            @Valid @RequestBody GroupMainPageLinkUpdateRequestDTO dto, @AuthenticationPrincipal AppUserDetails currentUser) {
        return ResponseEntity.ok(groupMainPageLinksService.update(id, dto, currentUser));
    }

    @PatchMapping("/safe-delete/{id}")
    @PreAuthorize("@security.hasAction('SAFE_DELETE_MAIN_PAGE_LINK') or @security.hasRole('ADMIN')")
    public ResponseEntity<Void> safeDelete(@PathVariable Integer id) {
        groupMainPageLinksService.safeDelete(id);
        return ResponseEntity.noContent().build();
    }

    @PatchMapping("/restore/{id}")
    @PreAuthorize("@security.hasAction('SAFE_DELETE_MAIN_PAGE_LINK') or @security.hasRole('ADMIN')")
    public ResponseEntity<Void> restore(@PathVariable Integer id) {
        groupMainPageLinksService.restore(id);
        return ResponseEntity.noContent().build();
    }

    @PreAuthorize("@security.hasAction('DELETE_MAIN_PAGE_LINK') or @security.hasRole('ADMIN')")
    @DeleteMapping("/{id}")
    public ResponseEntity<Void> permanentDelete(@PathVariable Integer id) {
        groupMainPageLinksService.permanentDelete(id);
        return ResponseEntity.noContent().build();
    }
}