# portal/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Портал Социальных аптек API
  version: 1.0.0
  description: |
    Контракт REST API портала. Аутентификация выполняется HTTP-only cookie `X-Access-Token`;
    обновление — cookie `X-Refresh-Token`. Для каждого POST, PUT, PATCH и DELETE передавайте
    cookie `XSRF-TOKEN` и одноимённый заголовок `X-XSRF-TOKEN`.
servers:
  - url: /
...
security:
  - cookieAuth: []
paths:
...
  /api/v1/tasks/{id}:
    get: { tags: [Задачи], summary: Получить задачу, parameters: [ { $ref: '#/components/parameters/TaskId' } ], responses: { '200': { $ref: '#/components/responses/Task' } } }
    put: { tags: [Задачи], summary: Изменить задачу, parameters: [ { $ref: '#/components/parameters/TaskId' }, { $ref: '#/components/parameters/CsrfHeader' } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/TaskUpdateRequest' } } } }, responses: { '200': { $ref: '#/components/responses/TaskShort' } } }
    delete: { tags: [Задачи], summary: Удалить задачу, parameters: [ { $ref: '#/components/parameters/TaskId' }, { $ref: '#/components/parameters/CsrfHeader' } ], responses: { '204': { description: Задача удалена } } }
...
  /api/v1/group-tasks:
    get: { tags: [Группы задач], summary: Получить группы задач по группам пользователей, parameters: [ { name: creatorGroupId, in: query, required: true, schema: { type: integer, minimum: 1 } }, { name: executorGroupId, in: query, required: true, schema: { type: integer, minimum: 1 } }, { $ref: '#/components/parameters/IsActive' } ], responses: { '200': { $ref: '#/components/responses/GroupTaskList' } } }
    post: { tags: [Группы задач], summary: Создать группу задач, parameters: [ { $ref: '#/components/parameters/CsrfHeader' } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/GroupTaskRequest' } } } }, responses: { '201': { $ref: '#/components/responses/GroupTask' } } }
  /api/v1/group-tasks/{id}:
    get: { tags: [Группы задач], summary: Получить группу задач, parameters: [ { $ref: '#/components/parameters/IntegerId' } ], responses: { '200': { $ref: '#/components/responses/GroupTask' } } }
    put: { tags: [Группы задач], summary: Обновить группу задач, parameters: [ { $ref: '#/components/parameters/IntegerId' }, { name: confirm, in: query, schema: { type: boolean, default: false } }, { $ref: '#/components/parameters/CsrfHeader' } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/GroupTaskUpdateRequest' } } } }, responses: { '200': { $ref: '#/components/responses/GroupTask' } } }
    delete: { tags: [Группы задач], summary: Удалить группу задач безвозвратно, parameters: [ { $ref: '#/components/parameters/IntegerId' }, { name: confirm, in: query, schema: { type: boolean, default: false } }, { $ref: '#/components/parameters/CsrfHeader' } ], responses: { '204': { description: Группа удалена } } }
  /api/v1/group-tasks/to/{executorGroupId}:
    get: { tags: [Группы задач], summary: Группы задач от моей группы, parameters: [ { name: executorGroupId, in: path, required: true, schema: { type: integer, minimum: 1 } }, { $ref: '#/components/parameters/IsActive' } ], responses: { '200': { $ref: '#/components/responses/GroupTaskList' } } }
  /api/v1/group-tasks/from/{creatorGroupId}:
    get: { tags: [Группы задач], summary: Группы задач для моей группы, parameters: [ { name: creatorGroupId, in: path, required: true, schema: { type: integer, minimum: 1 } }, { $ref: '#/components/parameters/IsActive' } ], responses: { '200': { $ref: '#/components/responses/GroupTaskList' } } }
  /api/v1/group-tasks/safe-delete/{id}:
    patch: { tags: [Группы задач], summary: Безопасно удалить группу задач, parameters: [ { $ref: '#/components/parameters/IntegerId' }, { $ref: '#/components/parameters/CsrfHeader' } ], responses: { '204': { description: Группа деактивирована } } }
  /api/v1/group-tasks/restore/{id}:
    patch: { tags: [Группы задач], summary: Восстановить группу задач, parameters: [ { $ref: '#/components/parameters/IntegerId' }, { $ref: '#/components/parameters/CsrfHeader' } ], responses: { '204': { description: Группа восстановлена } } }
...
  /api/v1/main-page-links/{id}:
    put: { tags: [Ссылки главной страницы], summary: Обновить ссылку, parameters: [ { $ref: '#/components/parameters/IntegerId' }, { $ref: '#/components/parameters/CsrfHeader' } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/MainPageLinkUpdateRequest' } } } }, responses: { '200': { $ref: '#/components/responses/MainPageLink' } } }
    delete: { tags: [Ссылки главной страницы], summary: Удалить ссылку, parameters: [ { $ref: '#/components/parameters/IntegerId' }, { $ref: '#/components/parameters/CsrfHeader' } ], responses: { '204': { description: Ссылка удалена } } }
  /api/v1/main-page-links/safe-delete/{id}:
    patch: { tags: [Ссылки главной страницы], summary: Безопасно удалить ссылку, parameters: [ { $ref: '#/components/parameters/IntegerId' }, { $ref: '#/components/parameters/CsrfHeader' } ], responses: { '204': { description: Ссылка деактивирована } } }
  /api/v1/main-page-links/restore/{id}:
    patch: { tags: [Ссылки главной страницы], summary: Восстановить ссылку, parameters: [ { $ref: '#/components/parameters/IntegerId' }, { $ref: '#/components/parameters/CsrfHeader' } ], responses: { '204': { description: Ссылка восстановлена } } }
  /api/v1/groups-main-page-links/get-all:
    get: { tags: [Ссылки главной страницы], summary: Получить группы ссылок, parameters: [ { $ref: '#/components/parameters/IsActive' } ], responses: { '200': { $ref: '#/components/responses/GroupMainPageLinksList' } } }
  /api/v1/groups-main-page-links:
    post: { tags: [Ссылки главной страницы], summary: Создать группу ссылок, parameters: [ { $ref: '#/components/parameters/CsrfHeader' } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/GroupMainPageLinksRequest' } } } }, responses: { '201': { $ref: '#/components/responses/GroupMainPageLinks' } } }
  /api/v1/groups-main-page-links/{id}:
    put: { tags: [Ссылки главной страницы], summary: Обновить группу ссылок, parameters: [ { $ref: '#/components/parameters/IntegerId' }, { $ref: '#/components/parameters/CsrfHeader' } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/GroupMainPageLinkUpdateRequest' } } } }, responses: { '200': { $ref: '#/components/responses/GroupMainPageLinks' } } }
    delete: { tags: [Ссылки главной страницы], summary: Удалить группу ссылок, parameters: [ { $ref: '#/components/parameters/IntegerId' }, { $ref: '#/components/parameters/CsrfHeader' } ], responses: { '204': { description: Группа ссылок удалена } } }
  /api/v1/groups-main-page-links/safe-delete/{id}:
    patch: { tags: [Ссылки главной страницы], summary: Безопасно удалить группу ссылок, parameters: [ { $ref: '#/components/parameters/IntegerId' }, { $ref: '#/components/parameters/CsrfHeader' } ], responses: { '204': { description: Группа ссылок деактивирована } } }
  /api/v1/groups-main-page-links/restore/{id}:
    patch: { tags: [Ссылки главной страницы], summary: Восстановить группу ссылок, parameters: [ { $ref: '#/components/parameters/IntegerId' }, { $ref: '#/components/parameters/CsrfHeader' } ], responses: { '204': { description: Группа ссылок восстановлена } } }
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: X-Access-Token
      description: JWT access token. Cookie `X-Refresh-Token` используется только `/api/v1/auth/refresh`.
    csrfToken:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
      description: Значение должно совпадать с cookie `XSRF-TOKEN`; обязательно для unsafe запросов.
  parameters:
    CsrfHeader: { name: X-XSRF-TOKEN, in: header, required: true, schema: { type: string }, description: Значение cookie XSRF-TOKEN }
...
    TaskId: { name: id, in: path, required: true, schema: { type: integer, format: int64, minimum: 1 } }
    IntegerId: { name: id, in: path, required: true, schema: { type: integer, minimum: 1 } }
...
    IsActive: { name: isActive, in: query, schema: { type: boolean, default: true } }
...
  responses:
...
    Task: { description: Задача, content: { application/json: { schema: { $ref: '#/components/schemas/Task' } } } }
    TaskShort: { description: Задача, content: { application/json: { schema: { $ref: '#/components/schemas/TaskShort' } } } }
...
    GroupTask: { description: Группа задач, content: { application/json: { schema: { $ref: '#/components/schemas/GroupTask' } } } }
    GroupTaskList: { description: Список групп задач, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/GroupTask' } } } } }
...
    MainPageLink: { description: Ссылка главной страницы, content: { application/json: { schema: { $ref: '#/components/schemas/MainPageLink' } } } }
...
    GroupMainPageLinks: { description: Группа ссылок, content: { application/json: { schema: { $ref: '#/components/schemas/GroupMainPageLinks' } } } }
    GroupMainPageLinksList: { description: Список групп ссылок, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/GroupMainPageLinks' } } } } }
  schemas:
...
    TaskStatus: { type: string, enum: [OPEN, CLOSED, DENIED, PROCESSED], description: 'OPEN — Открыта; CLOSED — Задача выполнена; DENIED — Задача отклонена; PROCESSED — В процессе выполнения.' }
    TaskPriority: { type: string, enum: [LOW, MIDDLE, HIGH], description: 'LOW — Низкий; MIDDLE — Средний; HIGH — Высокий.' }
...
    UserType: { type: string, enum: [CLIENT, APTEKA], description: 'CLIENT — Сотрудник; APTEKA — Аптека.' }
...
    TaskUpdateRequest: { type: object, properties: { title: { type: string, minLength: 1 }, description: { type: string, minLength: 1 }, workTypeId: { type: integer, minimum: 1 }, statusCode: { $ref: '#/components/schemas/TaskStatus' }, assignerId: { type: string, format: uuid } } }
...
    UserGroupShort: { type: object, properties: { id: { type: integer }, name: { type: string } } }
...
    UserShortInfo: { type: object, properties: { id: { type: string, format: uuid }, userType: { $ref: '#/components/schemas/UserType' }, displayName: { type: string } } }
    WorkType: { type: object, properties: { id: { type: integer }, name: { type: string }, wikiLink: { type: string, format: uri }, priorityResponseDTO: { type: object, properties: { priorityCode: { $ref: '#/components/schemas/TaskPriority' }, priorityDescription: { type: string } } }, taskGroup: { type: object, properties: { id: { type: integer }, name: { type: string } } } } }
    TaskPicture: { type: object, properties: { id: { type: integer, format: int64 }, downloadUrl: { type: string } } }
    TaskComment: { type: object, properties: { id: { type: integer, format: int64 }, comment: { type: string }, authorType: { $ref: '#/components/schemas/UserType' }, authorName: { type: string }, authorId: { nullable: true } } }
    TaskShort: { type: object, properties: { id: { type: integer, format: int64 }, title: { type: string }, description: { type: string }, creationDate: { type: string, format: date-time }, updatedDate: { type: string, format: date-time }, closingDate: { type: string, format: date-time, nullable: true }, status: { $ref: '#/components/schemas/TaskStatus' }, workType: { $ref: '#/components/schemas/WorkType' }, createdBy: { $ref: '#/components/schemas/UserShortInfo' }, assignedBy: { $ref: '#/components/schemas/UserShortInfo' } } }
    Task: { allOf: [ { $ref: '#/components/schemas/TaskShort' }, { type: object, properties: { comments: { type: array, items: { $ref: '#/components/schemas/TaskComment' } }, pictures: { type: array, items: { $ref: '#/components/schemas/TaskPicture' } } } } ] }
...
    GroupTaskRequest: { type: object, required: [name, creatorGroupId, intendedGroupId], properties: { name: { type: string, minLength: 1 }, creatorGroupId: { type: integer, minimum: 1 }, intendedGroupId: { type: integer, minimum: 1 } } }
    GroupTaskUpdateRequest: { type: object, required: [name], properties: { name: { type: string, minLength: 1 }, creatorGroupId: { type: integer, minimum: 1 }, intendedGroupId: { type: integer, minimum: 1 } } }
    GroupTask: { type: object, properties: { id: { type: integer }, name: { type: string }, creatorGroup: { $ref: '#/components/schemas/UserGroupShort' }, executorGroup: { $ref: '#/components/schemas/UserGroupShort' } } }
...
    MainPageLinkUpdateRequest: { type: object, additionalProperties: true, description: Запрос изменения ссылки главной страницы }
    MainPageLink: { type: object, additionalProperties: true, description: Ссылка главной страницы }
    GroupMainPageLinksRequest: { type: object, additionalProperties: true, description: Запрос создания группы ссылок }
    GroupMainPageLinkUpdateRequest: { type: object, additionalProperties: true, description: Запрос изменения группы ссылок }
    GroupMainPageLinks: { type: object, additionalProperties: true, description: Группа ссылок главной страницы }
...